Remove Jade from Services. Put getPostData in own file
Services will be variable in the future. I.e. the server implementor will choose
which services they want, and maybe write their own. This is a great reason for
service implementation to reside outside of the Services file. Additionaly,
this file was getting larger than I wanted it. There's no good reason (I can
think of) for individual services to be defined here.

This also gives us the opportunity to reload services without having to restart
the server when it comes to that (just invalidate the module cache).