Don't keep a revision history of node_modules.
package.json should be all we need to keep the revision history of.
Each machine will have to run npm update --save on it's own to get all the
node modules. Since some node modules have c code, they need to be compiled
per-machine

In order to do this, sadly I need to delete the whole folder and redownload