Replace plates with crispulent
Plates was giving me a bit of trouble. I ran into a bug, and I think it took me
less time to make crispulent which does what I need rather than read through
a few-hundred-line module and then try to find the bug and fix it.

Some terrible stuff went on in this commit. but I don't want to figure out how
to make crispulent into a module right now.. for now I'll just symlink it and
take the pain of doing it wrong so I can get more work done. This is a simple
and unforgettable fix.