Add parseAndValidateInput as a Utility
This sort of check will happen on almost all (if not all) post requests and
some or many get requests. Standardize it into a shared module instead of
defining these checks on a per-request basis.

There's some decent documentation on this funciton in the comments above it.
It definitely needs to be cleaned up a bit, but it's looking good.