Services: Add getSharedResponseObject
The comments are pretty in depth, but as a synopsis, this service is used
to get objects that are shared.

The object corresponds to the function which obtains that object, and
all attempts to get the object while the object is being gotten are added to
a queue which is cleared when the object has been loaded.